another step towards refactoring
