+++ django-backend/fecfiler/transactions/schedule_b/views.py
@@ -2,12 +2,12 @@ import logging
 
 from django.db.models.functions import Coalesce, Concat
 from datetime import datetime
-from rest_framework import filters
 from fecfiler.committee_accounts.views import CommitteeOwnedViewSet
 from fecfiler.f3x_summaries.views import ReportViewMixin
 
 from fecfiler.transactions.schedule_b.models import ScheduleBTransaction
-from .serializers import ScheduleATransactionSerializer
+from fecfiler.transactions.views import TransactionViewSetBase
+from fecfiler.transactions.schedule_b.serializers import ScheduleBTransactionSerializer
 from django.db.models import TextField, Value, Q
 from rest_framework.decorators import action
 from rest_framework.response import Response
@@ -16,7 +16,7 @@ from rest_framework import status
 logger = logging.getLogger(__name__)
 
 
-class ScheduleBTransactionViewSet(CommitteeOwnedViewSet, ReportViewMixin):
+class ScheduleBTransactionViewSet(TransactionViewSetBase):
     """
     This viewset automatically provides `list`, `create`, `retrieve`,
     `update` and `destroy` actions.
@@ -38,9 +38,7 @@ class ScheduleBTransactionViewSet(CommitteeOwnedViewSet, ReportViewMixin):
         )
     ).all()
 
-    serializer_class = ScheduleATransactionSerializer
-    permission_classes = []
-    filter_backends = [filters.OrderingFilter]
+    serializer_class = ScheduleBTransactionSerializer
     ordering_fields = [
         "id",
         "transaction_type_identifier",
@@ -49,7 +47,6 @@ class ScheduleBTransactionViewSet(CommitteeOwnedViewSet, ReportViewMixin):
         "memo_code",
         "expenditure_amount",
     ]
-    ordering = ["-created"]
 
     @action(detail=False, methods=["get"])
     def previous(self, request):
